add page for viewing team/member information

+++ src/main/java/com/example/smstest/domain/support/dto/SupportFilterCriteria.java
@@ -7,7 +7,7 @@ import java.util.List;
 @Data
 public class SupportFilterCriteria {
     private List<Long> customerId;
-    private List<Long> teamId;
+    private List<Integer> teamId;
     private List<Long> productId;
     private List<Long> issueId;
     private List<Long> stateId;
